requirements/base.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-jazzmin==2.6.0
django-split-settings==1.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
pip-chill==1.0.3
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
